+++ src/packets/GetEngineVersionPacket.js
@@ -41,17 +41,17 @@ GetEngineVersionPacket.prototype.write = function (writer) {
  * @param parser
  */
 GetEngineVersionPacket.prototype.parse = function (parser) {
-  var reponseLength = parser._parseInt();
+  var responseLength = parser._parseInt();
   this.casInfo = parser._parseBytes(DATA_TYPES.CAS_INFO_SIZE);
   this.responseCode = parser._parseInt();
   if (this.responseCode < 0) {
     this.errorCode = parser._parseInt();
-    this.errorMsg = parser._parseNullTerminatedString(reponseLength - 2 * DATA_TYPES.INT_SIZEOF);
+    this.errorMsg = parser._parseNullTerminatedString(responseLength - 2 * DATA_TYPES.INT_SIZEOF);
     if (this.errorMsg.length === 0) {
       this.errorMsg = Helpers._resolveErrorCode(this.errorCode);
     }
   } else {
-    this.engineVersion = parser._parseNullTerminatedString(reponseLength - DATA_TYPES.INT_SIZEOF); // Engine version
+    this.engineVersion = parser._parseNullTerminatedString(responseLength - DATA_TYPES.INT_SIZEOF); // Engine version
   }
 
   return this;