+++ src/packets/PrepareExecuteOldProtocolPacket.js
@@ -115,13 +115,13 @@ PrepareAndExecutePacket.prototype.writeExecute = function (writer) {
  * @param parser
  */
 PrepareAndExecutePacket.prototype.parsePrepare = function (parser) {
-  var reponseLength = parser._parseInt();
+  var responseLength = parser._parseInt();
   this.casInfo = parser._parseBytes(DATA_TYPES.CAS_INFO_SIZE);
 
   this.responseCode = parser._parseInt();
   if (this.responseCode < 0) {
     this.errorCode = parser._parseInt();
-    this.errorMsg = parser._parseNullTerminatedString(reponseLength - 2 * DATA_TYPES.INT_SIZEOF);
+    this.errorMsg = parser._parseNullTerminatedString(responseLength - 2 * DATA_TYPES.INT_SIZEOF);
     if (this.errorMsg.length === 0) {
       this.errorMsg = Helpers._resolveErrorCode(this.errorCode);
     }
@@ -165,13 +165,13 @@ PrepareAndExecutePacket.prototype.parsePrepare = function (parser) {
  * @param parser
  */
 PrepareAndExecutePacket.prototype.parseExecute = function (parser) {
-  var reponseLength = parser._parseInt();
+  var responseLength = parser._parseInt();
   this.casInfo = parser._parseBytes(DATA_TYPES.CAS_INFO_SIZE);
 
   this.responseCode = parser._parseInt();
   if (this.responseCode < 0) {
     this.errorCode = parser._parseInt();
-    this.errorMsg = parser._parseNullTerminatedString(reponseLength - 2 * DATA_TYPES.INT_SIZEOF);
+    this.errorMsg = parser._parseNullTerminatedString(responseLength - 2 * DATA_TYPES.INT_SIZEOF);
     if (this.errorMsg.length === 0) {
       this.errorMsg = Helpers._resolveErrorCode(this.errorCode);
     }