+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 # node-cubrid change log
 
-## Version 2.2.1 (Unreleased)
+## Version 2.2.1 (Feb 1, 2015)
 
 - Fix #15: When closing a query, if connection is closed, consider the request was successful.
+- Ref #14: Reduce the use of self where this can be used directly.
 - Travis: ohai requires Ruby version >= 2.0.0.
 
 ## Version 2.2.0 (May 15, 2014)